managed to fix scaling by using a smaller map that scales fairly well

+++ project-program/core/src/com/mygdx/game/PhysicsImp.java
@@ -10,7 +10,7 @@ public class PhysicsImp {
     public static final int S_HEIGHT = 759; // screeen height
     public static final int W_WIDTH = 10560 * 16; // world width
     public static final int W_HEIGHT = 50 * 16; // world height
-    public static final int UNITSCALE = 100; // this will scale up all our vector quantities to increase simulation speeds to something more realistic
+    public static final float UNITSCALE = 100; // this will scale up all our vector quantities to increase simulation speeds to something more realistic
     // setting up physical variables
     public static final float AIR_DENSITY = 1.225f;
